\name{fun1}
\alias{fun1}
\title{fun1}
\usage{
  fun1(x)
}
\arguments{
  \item{x}{numeric. this is multiplied by 100 to determine
  the length of the returned vector}
}
\value{
  a numeric vector of random deviates of length \code{100 *
  x}
}
\description{
  This is the Description section
}
\details{
  This is the Details section
}
\examples{
fun1(2)
length(fun1(20))
}
\author{
  your name
}
\seealso{
  \code{\link{fun2}}
}

